+++ CHANGELOG.md
@@ -20,6 +20,23 @@ Changelog
     bus to low / full speed, use `BusAdapter::with_speed`, and provide
     `Speed::LowFull`.
 
+- **BREAKING** Users now allocate space for endpoints. For backwards compatibility,
+  allocate the maximum amount of endpoints. Supply the endpoint state to your driver's
+  constructor, as show below.
+
+  ```rust
+  static EP_STATE: imxrt_usbd::EndpointState = imxrt_usbd::EndpointState::max_endpoints();
+
+  // ...
+
+  imxrt_usbd::BusAdapter::with_speed(
+      UsbPeripherals::usb1(),
+      &mut ENDPOINT_MEMORY,
+      &EP_STATE,                  // <-- NEW
+      SPEED,
+  )
+  ```
+
 - Add support for USB general purpose timers (GPT).
 - Fix the endpoint initialization routine, which would incorrectly zero the
   other half's endpoint type.